vpp-manager: refactor lifecycle, consolidate config
Restructure vpp-manager around an explicit VppRunner type and
move shared configuration into `pkg/config` so vpp-manager
and calico-vpp-agent reference a single source of truth.

- `pkg/config` is now the canonical home for VRF constants
  (DefaultVRFIndex, PuntTableID, PodVRFIndex), interface
  state/spec, machine state, env-var helpers, PingCalicoVpp(),
  ClearVppManagerFiles() and a single CreateVppLink(socket, log)
  used by both vpp-manager and calico-vpp-agent.
- `cmd/vpp-manager/main.go` is reduced to a thin entrypoint.
  Signal handling, uplink driver invocation and cleanup are
  centralized in `pkg/vpp-manager/vpp_runner.go` behind VppRunner,
  with deferred panic recovery that guarantees SIGINT to VPP and
  the VPP_ERRORED hook fires on any unexpected exit.
- Replaced embedded `default_hook.sh` and bash shell-out with a
  native Go NetworkManagerHook (`pkg/vpp-manager/hooks`).
  User-provided CALICOVPP_HOOK_* scripts override; otherwise the
  native hook runs when `CALICOVPP_ENABLE_NETWORK_MANAGER_HOOK=true`
  (default), and is a no-op when disabled.
- Removed SupportedUplinkDrivers auto-detect retry; uplink driver
  is chosen deterministically from the configured interface spec.
- All operational loops over uplinks iterate `InterfacesByID`
  (ordered slice) instead of `Interfaces` (Go map) so template
  rendering and driver step ordering is deterministic run-to-run.
