+++ pkg/vpp-manager/hooks/network_manager_hook.go
@@ -26,6 +26,7 @@ import (
 
 	"github.com/pkg/errors"
 	"github.com/projectcalico/vpp-dataplane/v3/pkg/config"
+	"github.com/projectcalico/vpp-dataplane/v3/pkg/vpp-manager/params"
 	"github.com/sirupsen/logrus"
 )
 
@@ -64,10 +65,11 @@ type UdevNetNameProperties struct {
 
 // NetworkManagerHook manages network configuration during VPP lifecycle
 type NetworkManagerHook struct {
-	interfaceNames []string
-	systemType     SystemType
-	log            *logrus.Logger
-	udevProps      map[string]*UdevNetNameProperties // map[interfaceName] -> udev properties
+	interfaceNames   []string
+	systemType       SystemType
+	log              *logrus.Entry
+	udevProps        map[string]*UdevNetNameProperties // map[interfaceName] -> udev properties
+	vppManagerParams *params.VppManagerParams
 }
 
 // chrootCommand creates a command that will be executed in the host namespace
@@ -186,20 +188,15 @@ func (h *NetworkManagerHook) restartSystemdNetworkd() error {
 }
 
 // NewNetworkManagerHook creates a new NetworkManagerHook instance
-func NewNetworkManagerHook(log *logrus.Logger) *NetworkManagerHook {
+func NewNetworkManagerHook(log *logrus.Entry, vppManagerParams *params.VppManagerParams) *NetworkManagerHook {
 	hook := &NetworkManagerHook{
-		log:       log,
-		udevProps: make(map[string]*UdevNetNameProperties),
+		log:              log,
+		udevProps:        make(map[string]*UdevNetNameProperties),
+		vppManagerParams: vppManagerParams,
 	}
 
 	hook.detectSystem()
 	return hook
-}
-
-// SetInterfaceNames updates the interface names of NetworkManagerHook instance
-func (h *NetworkManagerHook) SetInterfaceNames(interfaceNames []string) {
-	h.interfaceNames = interfaceNames
-	h.log.Infof("NetworkManagerHook: Interface names updated to %v", interfaceNames)
 }
 
 // fixDNS modifies NetworkManager configuration to disable DNS management
@@ -730,8 +727,8 @@ func (h *NetworkManagerHook) beforeVppRun() error {
 	}
 
 	// Save network file for AWS systemd-networkd for each interface
-	for _, interfaceName := range h.interfaceNames {
-		err = h.saveNetworkFile(interfaceName)
+	for _, intf := range h.vppManagerParams.InterfacesByID {
+		err = h.saveNetworkFile(intf.Spec.InterfaceName)
 		if err != nil {
 			return err
 		}
@@ -749,8 +746,8 @@ func (h *NetworkManagerHook) vppRunning() error {
 	}
 
 	// Tweak network file for AWS systemd-networkd for each interface
-	for _, interfaceName := range h.interfaceNames {
-		err = h.tweakNetworkFile(interfaceName)
+	for _, intf := range h.vppManagerParams.InterfacesByID {
+		err = h.tweakNetworkFile(intf.Spec.InterfaceName)
 		if err != nil {
 			return err
 		}
@@ -774,8 +771,8 @@ func (h *NetworkManagerHook) vppDoneOk() error {
 	}
 
 	// Remove the tweaked network file for AWS systemd-networkd for each interface
-	for _, interfaceName := range h.interfaceNames {
-		err = h.removeTweakedNetworkFile(interfaceName)
+	for _, intf := range h.vppManagerParams.InterfacesByID {
+		err = h.removeTweakedNetworkFile(intf.Spec.InterfaceName)
 		if err != nil {
 			return err
 		}
@@ -802,7 +799,7 @@ func (h *NetworkManagerHook) Execute(hookPoint HookPoint) error {
 		return nil
 	}
 
-	h.log.Infof("NetworkManagerHook: Executing %s for interfaces %v", hookPoint, h.interfaceNames)
+	h.log.Infof("NetworkManagerHook: Executing %s for all interfaces", hookPoint)
 
 	var err error
 	switch hookPoint {
@@ -837,26 +834,23 @@ func (h *NetworkManagerHook) runGoNativeHook(userHook *string) bool {
 	return userHook == nil || *userHook == ""
 }
 
-// useDefaultHookFallback returns true when we should use the legacy default_hook.sh
-// This happens when: native hooks are disabled AND no user script is configured
-func (h *NetworkManagerHook) useDefaultHookFallback(userHook *string) bool {
-	if config.EnableNetworkManagerHook != nil && *config.EnableNetworkManagerHook {
-		return false // Native hooks enabled, no fallback needed
-	}
-	// Native hooks disabled, use fallback only if no user script
-	return userHook == nil || *userHook == ""
-}
-
 // ExecuteWithUserScript runs the appropriate hook based on configuration:
 // 1. If user script configured -> run user script only (override)
 // 2. If native hooks enabled -> run native Go hook
-// 3. If native hooks disabled & no user script -> fallback to default_hook.sh
+// 3. If native hooks disabled & no user script -> no-op
 func (h *NetworkManagerHook) ExecuteWithUserScript(hookPoint HookPoint,
-	hookScript *string,
-	params *config.VppManagerParams) {
+	hookScript *string) {
 	// Check if user script is configured (highest priority - overrides everything)
 	if hookScript != nil && *hookScript != "" {
-		config.RunHook(hookScript, string(hookPoint), params, h.log)
+		template := h.vppManagerParams.TemplateScriptReplace(*hookScript)
+
+		cmd := exec.Command("/bin/bash", "-c", template, string(hookPoint), h.vppManagerParams.InterfacesByID[0].Spec.InterfaceName)
+		cmd.Stdout = os.Stdout
+		cmd.Stderr = os.Stderr
+		err := cmd.Run()
+		if err != nil {
+			h.log.Warnf("Running hook %s errored with %s", string(hookPoint), err)
+		}
 		return
 	}
 
@@ -867,11 +861,5 @@ func (h *NetworkManagerHook) ExecuteWithUserScript(hookPoint HookPoint,
 			h.log.Warnf("Network hook %s failed: %v", hookPoint, err)
 		}
 		return
-	}
-
-	// Fallback to default_hook.sh when native hooks are disabled
-	if h.useDefaultHookFallback(hookScript) {
-		h.log.Infof("Native hooks disabled, falling back to default_hook.sh for %s", hookPoint)
-		config.RunHook(&config.DefaultHookScript, string(hookPoint), params, h.log)
 	}
 }
