--- pkg/vpp-manager/startup/interface_config.go
@@ -1,152 +1,0 @@
-// Copyright (C) 2019 Cisco Systems Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//    http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
-// implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package startup
-
-import (
-	"encoding/gob"
-	"fmt"
-	"os"
-
-	"github.com/pkg/errors"
-	log "github.com/sirupsen/logrus"
-
-	"github.com/projectcalico/vpp-dataplane/v3/pkg/config"
-	"github.com/projectcalico/vpp-dataplane/v3/pkg/vpp-manager/utils"
-)
-
-func GetInterfaceConfig(params *config.VppManagerParams) (conf []*config.LinuxInterfaceState, err error) {
-	errs := []error{}
-	conf = []*config.LinuxInterfaceState{}
-	for _, ifSpec := range params.UplinksSpecs {
-		configuration, err := loadInterfaceConfigFromLinux(ifSpec)
-		errs = append(errs, err)
-		conf = append(conf, configuration)
-	}
-	allLoaded := true
-	for i := range errs {
-		if errs[i] != nil {
-			allLoaded = false
-			log.Warnf("Could not load config from linux (%v)", errs[i])
-		}
-	}
-	if allLoaded {
-		err = saveConfig(params, conf)
-		if err != nil {
-			log.Warnf("Could not save interface config: %v", err)
-		}
-	} else {
-		// Loading config failed, try loading from save file
-		confFile, err2 := loadInterfaceConfigFromFile(params)
-		if err2 != nil {
-			return nil, err2
-		}
-		// If loaded from file replace non loaded interface configs
-		for i := range conf {
-			if conf[i] == nil {
-				for j := range confFile {
-					if confFile[j].InterfaceName == params.UplinksSpecs[i].InterfaceName {
-						conf[i] = confFile[j]
-					}
-				}
-			}
-		}
-		for i := range conf {
-			if conf[i] == nil {
-				return nil, fmt.Errorf("interface configs not found")
-			}
-		}
-		log.Infof("Loaded config. Interfaces marked as down since loading config from linux failed.")
-		// This ensures we don't try to set the interface down in runVpp()
-		for _, config := range conf {
-			config.IsUp = false
-		}
-	}
-	return conf, nil
-}
-
-func loadInterfaceConfigFromLinux(ifSpec config.UplinkInterfaceSpec) (*config.LinuxInterfaceState, error) {
-	conf, err := config.LoadInterfaceConfigFromLinux(ifSpec.InterfaceName, ifSpec.IPFamilies)
-	if err != nil {
-		return nil, err
-	}
-
-	pciID, err := utils.GetInterfacePciID(ifSpec.InterfaceName)
-	// We allow PCI not to be found e.g for AF_PACKET
-	if err != nil || pciID == "" {
-		log.Infof("No pci device for interface %s", ifSpec.InterfaceName)
-	} else {
-		conf.PciID = pciID
-		driver, err := utils.GetDriverNameFromPci(pciID)
-		if err != nil {
-			return nil, err
-		}
-		conf.Driver = driver
-		if ifSpec.NewDriverName != "" && ifSpec.NewDriverName != conf.Driver {
-			conf.DoSwapDriver = true
-		}
-	}
-	return conf, nil
-}
-
-func clearSavedConfig(params *config.VppManagerParams) {
-	if config.GetCalicoVppInitialConfig().IfConfigSavePath == "" {
-		return
-	}
-	err := os.Remove(config.GetCalicoVppInitialConfig().IfConfigSavePath)
-	if err != nil {
-		log.Warnf("could not delete saved interface config: %v", err)
-	}
-}
-
-func saveConfig(params *config.VppManagerParams, conf []*config.LinuxInterfaceState) error {
-	if config.GetCalicoVppInitialConfig().IfConfigSavePath == "" {
-		return nil
-	}
-	file, err := os.Create(config.GetCalicoVppInitialConfig().IfConfigSavePath)
-	if err != nil {
-		return errors.Wrap(err, "error opening save file")
-	}
-	enc := gob.NewEncoder(file)
-	err = enc.Encode(conf)
-	if err != nil {
-		file.Close()
-		clearSavedConfig(params)
-		return errors.Wrap(err, "error encoding data")
-	}
-	err = file.Close()
-	if err != nil {
-		return errors.Wrap(err, "error closing file")
-	}
-	return nil
-}
-
-func loadInterfaceConfigFromFile(params *config.VppManagerParams) ([]*config.LinuxInterfaceState, error) {
-	conf := []*config.LinuxInterfaceState{}
-	if config.GetCalicoVppInitialConfig().IfConfigSavePath == "" {
-		return nil, fmt.Errorf("interface config save file not configured")
-	}
-	file, err := os.Open(config.GetCalicoVppInitialConfig().IfConfigSavePath)
-	if err != nil {
-		return nil, errors.Wrap(err, "error opening save file")
-	}
-	dec := gob.NewDecoder(file)
-	err = dec.Decode(&conf)
-	if err != nil {
-		return nil, errors.Wrap(err, "decode error")
-	}
-	file.Close()
-	return conf, nil
-}
