+++ pkg/vpp-manager/uplink/dpdk.go
@@ -22,11 +22,11 @@ import (
 	"time"
 
 	"github.com/pkg/errors"
-	log "github.com/sirupsen/logrus"
+	"github.com/sirupsen/logrus"
 	"github.com/vishvananda/netlink"
 
 	"github.com/projectcalico/vpp-dataplane/v3/pkg/config"
-	"github.com/projectcalico/vpp-dataplane/v3/pkg/vpp-manager/utils"
+	vppmanagerparams "github.com/projectcalico/vpp-dataplane/v3/pkg/vpp-manager/params"
 	"github.com/projectcalico/vpp-dataplane/v3/pkg/vpplink"
 	"github.com/projectcalico/vpp-dataplane/v3/pkg/vpplink/types"
 )
@@ -38,33 +38,33 @@ type DPDKDriver struct {
 func (d *DPDKDriver) IsSupported(warn bool) (supported bool) {
 	var ret bool
 	supported = true
-	ret = d.conf.PciID != ""
+	ret = d.intf.State.PciID != ""
 	if !ret && warn {
-		log.Warnf("did not find  pci device id for interface")
+		d.log.Warnf("did not find  pci device id for interface")
 	}
 	supported = supported && ret
 	return supported
 }
 
 func (d *DPDKDriver) getFinalDriver() string {
-	if d.conf.DoSwapDriver {
-		return d.spec.NewDriverName
+	if d.intf.State.DoSwapDriver {
+		return d.intf.Spec.NewDriverName
 	}
-	return d.conf.Driver
+	return d.intf.State.Driver
 }
 
 func (d *DPDKDriver) PreconfigureLinux() (err error) {
 	d.removeLinuxIfConf(true /* down */)
-	if d.conf.DoSwapDriver {
-		err = utils.SwapDriver(d.conf.PciID, d.spec.NewDriverName, true)
+	if d.intf.State.DoSwapDriver {
+		err = config.SwapDriver(d.log, d.intf.State.PciID, d.intf.Spec.NewDriverName, true)
 		if err != nil {
-			log.Warnf("Failed to swap driver to %s: %v", d.spec.NewDriverName, err)
+			d.log.Warnf("Failed to swap driver to %s: %v", d.intf.Spec.NewDriverName, err)
 		}
 	}
 	if d.getFinalDriver() == config.DriverVfioPci &&
 		d.params.AvailableHugePages == 0 &&
 		d.params.InitialVfioEnableUnsafeNoIommuMode == config.VfioUnsafeNoIommuModeYES {
-		err := utils.SetVfioEnableUnsafeNoIommuMode(config.VfioUnsafeNoIommuModeNO)
+		err := config.SetVfioEnableUnsafeNoIommuMode(config.VfioUnsafeNoIommuModeNO)
 		if err != nil {
 			return errors.Wrapf(err, "failed to configure vfio")
 		}
@@ -84,15 +84,15 @@ func (d *DPDKDriver) UpdateVppConfigFile(template string) string {
 	if d.params.AvailableHugePages > 0 {
 		template = fmt.Sprintf(
 			"%s\ndpdk {\ndev %s { num-rx-queues %d num-tx-queues %d num-rx-desc %d num-tx-desc %d tag %s } \n}\n",
-			template, d.conf.PciID, d.spec.NumRxQueues, d.spec.NumTxQueues,
-			d.spec.RxQueueSize, d.spec.TxQueueSize, "main-"+d.spec.InterfaceName,
+			template, d.intf.State.PciID, d.intf.Spec.NumRxQueues, d.intf.Spec.NumTxQueues,
+			d.intf.Spec.RxQueueSize, d.intf.Spec.TxQueueSize, "main-"+d.intf.Spec.InterfaceName,
 		)
 
 	} else {
 		template = fmt.Sprintf(
 			"%s\ndpdk {\niova-mode va\nno-hugetlb\ndev %s { num-rx-queues %d num-tx-queues %d num-rx-desc %d num-tx-desc %d tag %s } \n}\n",
-			template, d.conf.PciID, d.spec.NumRxQueues, d.spec.NumTxQueues,
-			d.spec.RxQueueSize, d.spec.TxQueueSize, "main-"+d.spec.InterfaceName,
+			template, d.intf.State.PciID, d.intf.Spec.NumRxQueues, d.intf.Spec.NumTxQueues,
+			d.intf.Spec.RxQueueSize, d.intf.Spec.TxQueueSize, "main-"+d.intf.Spec.InterfaceName,
 		)
 
 		// If no hugepages, also edit `buffers {}`
@@ -102,7 +102,7 @@ func (d *DPDKDriver) UpdateVppConfigFile(template string) string {
 		if buffersStanzaRegex.MatchString(template) {
 			if !buffersNoHugeStanzaRegex.MatchString(template) {
 				template = buffersHeadRegex.ReplaceAllString(template, "buffers {\nno-hugetlb")
-				log.Infof("Found buffers configuration in template")
+				d.log.Infof("Found buffers configuration in template")
 			}
 		} else {
 			template = fmt.Sprintf("%s\nbuffers {\nno-hugetlb\n}", template)
@@ -114,50 +114,50 @@ write:
 }
 
 func (d *DPDKDriver) restoreInterfaceName() error {
-	newName, err := utils.GetInterfaceNameFromPci(d.conf.PciID)
+	newName, err := config.GetInterfaceNameFromPci(d.intf.State.PciID)
 	if err != nil {
-		return errors.Wrapf(err, "Error getting new if name for %s: %v", newName, d.conf.PciID)
+		return errors.Wrapf(err, "Error getting new if name for %s: %v", newName, d.intf.State.PciID)
 	}
-	if newName == d.spec.InterfaceName {
+	if newName == d.intf.Spec.InterfaceName {
 		return nil
 	}
 	link, err := netlink.LinkByName(newName)
 	if err != nil {
 		return errors.Wrapf(err, "Error getting new link %s: %v", newName, link)
 	}
-	err = netlink.LinkSetName(link, d.spec.InterfaceName)
+	err = netlink.LinkSetName(link, d.intf.Spec.InterfaceName)
 	if err != nil {
-		return errors.Wrapf(err, "Error setting new if name for %s: %v", d.spec.InterfaceName, link)
+		return errors.Wrapf(err, "Error setting new if name for %s: %v", d.intf.Spec.InterfaceName, link)
 	}
 	return nil
 }
 
-func (d *DPDKDriver) RestoreLinux(allInterfacesPhysical bool) {
-	if d.conf.PciID != "" && d.conf.Driver != "" {
-		err := utils.SwapDriver(d.conf.PciID, d.conf.Driver, false)
+func (d *DPDKDriver) RestoreLinux() {
+	if d.intf.State.PciID != "" && d.intf.State.Driver != "" {
+		err := config.SwapDriver(d.log, d.intf.State.PciID, d.intf.State.Driver, false)
 		if err != nil {
-			log.Warnf("Error swapping back driver to %s for %s: %v", d.conf.Driver, d.conf.PciID, err)
+			d.log.Warnf("Error swapping back driver to %s for %s: %v", d.intf.State.Driver, d.intf.State.PciID, err)
 		}
 	}
 
 	for i := 0; i < 10; i++ {
 		err := d.restoreInterfaceName()
 		if err != nil {
-			log.Warnf("Error restoring if name %s", err)
+			d.log.Warnf("Error restoring if name %s", err)
 		} else {
 			break
 		}
 		time.Sleep(500 * time.Millisecond)
 	}
 
-	if !d.conf.IsUp {
+	if !d.intf.State.IsUp {
 		return
 	}
 	// This assumes the link has kept the same name after the rebind.
 	// It should be always true on systemd based distros
-	link, err := utils.SafeSetInterfaceUpByName(d.spec.InterfaceName)
+	link, err := config.SafeSetInterfaceUpByName(d.intf.Spec.InterfaceName)
 	if err != nil {
-		log.Warnf("Error setting %s up: %v", d.spec.InterfaceName, err)
+		d.log.Warnf("Error setting %s up: %v", d.intf.Spec.InterfaceName, err)
 		return
 	}
 
@@ -167,9 +167,9 @@ func (d *DPDKDriver) RestoreLinux(allInterfacesPhysical bool) {
 	if d.getFinalDriver() == config.DriverVfioPci &&
 		d.params.AvailableHugePages == 0 &&
 		d.params.InitialVfioEnableUnsafeNoIommuMode == config.VfioUnsafeNoIommuModeYES {
-		err = utils.SetVfioEnableUnsafeNoIommuMode(config.VfioUnsafeNoIommuModeYES)
+		err = config.SetVfioEnableUnsafeNoIommuMode(config.VfioUnsafeNoIommuModeYES)
 		if err != nil {
-			log.Errorf("failed to configure vfio %s", err)
+			d.log.Errorf("failed to configure vfio %s", err)
 		}
 	}
 }
@@ -177,29 +177,33 @@ func (d *DPDKDriver) RestoreLinux(allInterfacesPhysical bool) {
 func (d *DPDKDriver) CreateMainVppInterface(vpp *vpplink.VppLink, vppPid int, uplinkSpec *config.UplinkInterfaceSpec) (err error) {
 	// Nothing to do VPP autocreates on startup
 	// refusing to run on secondary interfaces as we have no way to figure out the sw_if_index
-	if !d.spec.IsMain {
+	if !d.intf.Spec.IsMain {
 		return fmt.Errorf("%s driver not supported for secondary interfaces", d.name)
 	}
-	swIfIndex, err := vpp.SearchInterfaceWithTag("main-" + d.spec.InterfaceName)
+	swIfIndex, err := vpp.SearchInterfaceWithTag("main-" + d.intf.Spec.InterfaceName)
 	if err != nil || swIfIndex == ^uint32(0) {
-		return fmt.Errorf("error trying to find interface with tag main-%s", d.spec.InterfaceName)
+		return fmt.Errorf("error trying to find interface with tag main-%s", d.intf.Spec.InterfaceName)
 	}
-	log.Debugf("Found interface with swIfIndex %d for %s", swIfIndex, d.spec.InterfaceName)
-	d.spec.SwIfIndex = swIfIndex
-	err = vpp.SetInterfaceMacAddress(swIfIndex, d.conf.HardwareAddr)
+	d.log.Debugf("Found interface with swIfIndex %d for %s", swIfIndex, d.intf.Spec.InterfaceName)
+	d.intf.Spec.SwIfIndex = swIfIndex
+	err = vpp.SetInterfaceMacAddress(swIfIndex, d.intf.State.HardwareAddr)
 	if err != nil && gerrors.Is(err, types.VppErrorUnimplemented) {
-		log.Warn("Setting dpdk interface mac address in vpp unsupported")
+		d.log.Warn("Setting dpdk interface mac address in vpp unsupported")
 	} else if err != nil {
 		return errors.Wrapf(err, "could not set dpdk interface %d mac address in vpp", swIfIndex)
 	}
 	return nil
 }
 
-func NewDPDKDriver(params *config.VppManagerParams, conf *config.LinuxInterfaceState, spec *config.UplinkInterfaceSpec) *DPDKDriver {
-	d := &DPDKDriver{}
-	d.name = NativeDriverDpdk
-	d.conf = conf
-	d.params = params
-	d.spec = spec
-	return d
+func NewDPDKDriver(params *vppmanagerparams.VppManagerParams,
+	intf *vppmanagerparams.VppManagerInterface,
+	log *logrus.Entry) *DPDKDriver {
+	return &DPDKDriver{
+		UplinkDriverData: UplinkDriverData{
+			name:   NativeDriverDpdk,
+			params: params,
+			intf:   intf,
+			log:    log,
+		},
+	}
 }
