+++ pkg/vpp-manager/uplink/virtio.go
@@ -17,10 +17,10 @@ package uplink
 
 import (
 	"github.com/pkg/errors"
-	log "github.com/sirupsen/logrus"
+	"github.com/sirupsen/logrus"
 
 	"github.com/projectcalico/vpp-dataplane/v3/pkg/config"
-	"github.com/projectcalico/vpp-dataplane/v3/pkg/vpp-manager/utils"
+	vppmanagerparams "github.com/projectcalico/vpp-dataplane/v3/pkg/vpp-manager/params"
 	"github.com/projectcalico/vpp-dataplane/v3/pkg/vpplink"
 	"github.com/projectcalico/vpp-dataplane/v3/pkg/vpplink/types"
 )
@@ -34,20 +34,20 @@ func (d *VirtioDriver) IsSupported(warn bool) (supported bool) {
 	supported = true
 	ret = d.params.LoadedDrivers[config.DriverVfioPci]
 	if !ret && warn {
-		log.Warnf("did not find vfio-pci or uio_pci_generic driver")
-		log.Warnf("VPP may fail to grab its interface")
+		d.log.Warnf("did not find vfio-pci or uio_pci_generic driver")
+		d.log.Warnf("VPP may fail to grab its interface")
 	}
 	supported = supported && ret
 
 	ret = d.params.AvailableHugePages > 0
 	if !ret && warn {
-		log.Warnf("No hugepages configured, driver won't work")
+		d.log.Warnf("No hugepages configured, driver won't work")
 	}
 	supported = supported && ret
 
-	ret = d.conf.Driver == config.DriverVirtioPci
+	ret = d.intf.State.Driver == config.DriverVirtioPci
 	if !ret && warn {
-		log.Warnf("Interface driver is <%s>, not %s", d.conf.Driver, config.DriverVirtioPci)
+		d.log.Warnf("Interface driver is <%s>, not %s", d.intf.State.Driver, config.DriverVirtioPci)
 	}
 	supported = supported && ret
 
@@ -55,56 +55,56 @@ func (d *VirtioDriver) IsSupported(warn bool) (supported bool) {
 }
 
 func (d *VirtioDriver) PreconfigureLinux() (err error) {
-	newDriverName := d.spec.NewDriverName
-	doSwapDriver := d.conf.DoSwapDriver
+	newDriverName := d.intf.Spec.NewDriverName
+	doSwapDriver := d.intf.State.DoSwapDriver
 	if newDriverName == "" {
 		newDriverName = config.DriverVfioPci
-		doSwapDriver = config.DriverVfioPci != d.conf.Driver
+		doSwapDriver = config.DriverVfioPci != d.intf.State.Driver
 	}
 
 	if d.params.InitialVfioEnableUnsafeNoIommuMode == config.VfioUnsafeNoIommuModeNO {
-		err := utils.SetVfioEnableUnsafeNoIommuMode(config.VfioUnsafeNoIommuModeYES)
+		err := config.SetVfioEnableUnsafeNoIommuMode(config.VfioUnsafeNoIommuModeYES)
 		if err != nil {
 			return errors.Wrapf(err, "failed to configure vfio")
 		}
 	}
 	d.removeLinuxIfConf(true /* down */)
 	if doSwapDriver {
-		err = utils.SwapDriver(d.conf.PciID, newDriverName, true)
+		err = config.SwapDriver(d.log, d.intf.State.PciID, newDriverName, true)
 		if err != nil {
-			log.Warnf("Failed to swap driver to %s: %v", newDriverName, err)
+			d.log.Warnf("Failed to swap driver to %s: %v", newDriverName, err)
 		}
 	}
 	return nil
 }
 
-func (d *VirtioDriver) RestoreLinux(allInterfacesPhysical bool) {
+func (d *VirtioDriver) RestoreLinux() {
 	if d.params.InitialVfioEnableUnsafeNoIommuMode == config.VfioUnsafeNoIommuModeNO {
-		err := utils.SetVfioEnableUnsafeNoIommuMode(config.VfioUnsafeNoIommuModeNO)
+		err := config.SetVfioEnableUnsafeNoIommuMode(config.VfioUnsafeNoIommuModeNO)
 		if err != nil {
-			log.Warnf("Virtio restore error %v", err)
+			d.log.Warnf("Virtio restore error %v", err)
 		}
 	}
-	if d.conf.PciID != "" && d.conf.Driver != "" {
-		err := utils.SwapDriver(d.conf.PciID, d.conf.Driver, false)
+	if d.intf.State.PciID != "" && d.intf.State.Driver != "" {
+		err := config.SwapDriver(d.log, d.intf.State.PciID, d.intf.State.Driver, false)
 		if err != nil {
-			log.Warnf("Error swapping back driver to %s for %s: %v", d.conf.Driver, d.conf.PciID, err)
+			d.log.Warnf("Error swapping back driver to %s for %s: %v", d.intf.State.Driver, d.intf.State.PciID, err)
 		}
 	}
-	if !allInterfacesPhysical {
-		err := d.moveInterfaceFromNS(d.spec.InterfaceName)
+	if !d.params.AllInterfacesPhysical() {
+		err := d.moveInterfaceFromNS(d.intf.Spec.InterfaceName)
 		if err != nil {
-			log.Warnf("Moving uplink back from NS failed %s", err)
+			d.log.Warnf("Moving uplink back from NS failed %s", err)
 		}
 	}
-	if !d.conf.IsUp {
+	if !d.intf.State.IsUp {
 		return
 	}
 	// This assumes the link has kept the same name after the rebind.
 	// It should be always true on systemd based distros
-	link, err := utils.SafeSetInterfaceUpByName(d.spec.InterfaceName)
+	link, err := config.SafeSetInterfaceUpByName(d.intf.Spec.InterfaceName)
 	if err != nil {
-		log.Warnf("Error setting %s up: %v", d.spec.InterfaceName, err)
+		d.log.Warnf("Error setting %s up: %v", d.intf.Spec.InterfaceName, err)
 		return
 	}
 
@@ -115,27 +115,31 @@ func (d *VirtioDriver) RestoreLinux(allInterfacesPhysical bool) {
 func (d *VirtioDriver) CreateMainVppInterface(vpp *vpplink.VppLink, vppPid int, uplinkSpec *config.UplinkInterfaceSpec) (err error) {
 	intf := types.VirtioInterface{
 		GenericVppInterface: d.getGenericVppInterface(),
-		PciID:               d.conf.PciID,
+		PciID:               d.intf.State.PciID,
 	}
 	swIfIndex, err := vpp.CreateVirtio(&intf)
 	if err != nil {
 		return errors.Wrapf(err, "Error creating VIRTIO interface")
 	}
-	log.Infof("Created VIRTIO interface %d", swIfIndex)
+	d.log.Infof("Created VIRTIO interface %d", swIfIndex)
 
-	d.spec.SwIfIndex = swIfIndex
-	err = d.TagMainInterface(vpp, swIfIndex, d.spec.InterfaceName)
+	d.intf.Spec.SwIfIndex = swIfIndex
+	err = d.TagMainInterface(vpp, swIfIndex, d.intf.Spec.InterfaceName)
 	if err != nil {
 		return err
 	}
 	return nil
 }
 
-func NewVirtioDriver(params *config.VppManagerParams, conf *config.LinuxInterfaceState, spec *config.UplinkInterfaceSpec) *VirtioDriver {
-	d := &VirtioDriver{}
-	d.name = NativeDriverVirtio
-	d.conf = conf
-	d.params = params
-	d.spec = spec
-	return d
+func NewVirtioDriver(params *vppmanagerparams.VppManagerParams,
+	intf *vppmanagerparams.VppManagerInterface,
+	log *logrus.Entry) *VirtioDriver {
+	return &VirtioDriver{
+		UplinkDriverData: UplinkDriverData{
+			name:   NativeDriverVirtio,
+			params: params,
+			intf:   intf,
+			log:    log,
+		},
+	}
 }
