+++ pkg/vpp-manager/uplink/vmxnet3.go
@@ -17,10 +17,10 @@ package uplink
 
 import (
 	"github.com/pkg/errors"
-	log "github.com/sirupsen/logrus"
+	"github.com/sirupsen/logrus"
 
 	"github.com/projectcalico/vpp-dataplane/v3/pkg/config"
-	"github.com/projectcalico/vpp-dataplane/v3/pkg/vpp-manager/utils"
+	vppmanagerparams "github.com/projectcalico/vpp-dataplane/v3/pkg/vpp-manager/params"
 	"github.com/projectcalico/vpp-dataplane/v3/pkg/vpplink"
 	"github.com/projectcalico/vpp-dataplane/v3/pkg/vpplink/types"
 )
@@ -33,9 +33,9 @@ func (d *Vmxnet3Driver) IsSupported(warn bool) (supported bool) {
 	var ret bool
 	supported = true
 
-	ret = d.conf.Driver == config.DriverVmxNet3
+	ret = d.intf.State.Driver == config.DriverVmxNet3
 	if !ret && warn {
-		log.Warnf("Interface driver is <%s>, not %s", d.conf.Driver, config.DriverVmxNet3)
+		d.log.Warnf("Interface driver is <%s>, not %s", d.intf.State.Driver, config.DriverVmxNet3)
 	}
 	supported = supported && ret
 
@@ -44,41 +44,41 @@ func (d *Vmxnet3Driver) IsSupported(warn bool) (supported bool) {
 
 func (d *Vmxnet3Driver) PreconfigureLinux() (err error) {
 	if d.params.InitialVfioEnableUnsafeNoIommuMode == config.VfioUnsafeNoIommuModeNO {
-		err := utils.SetVfioEnableUnsafeNoIommuMode(config.VfioUnsafeNoIommuModeYES)
+		err := config.SetVfioEnableUnsafeNoIommuMode(config.VfioUnsafeNoIommuModeYES)
 		if err != nil {
 			return errors.Wrapf(err, "Vmxnet3 preconfigure error")
 		}
 	}
 	d.removeLinuxIfConf(true /* down */)
-	driverName, err := utils.GetDriverNameFromPci(d.conf.PciID)
+	driverName, err := config.GetDriverNameFromPci(d.intf.State.PciID)
 	if err != nil {
-		return errors.Wrapf(err, "Couldnt get VF driver Name for %s", d.conf.PciID)
+		return errors.Wrapf(err, "Couldnt get VF driver Name for %s", d.intf.State.PciID)
 	}
 	if driverName != config.DriverVfioPci {
-		err := utils.SwapDriver(d.conf.PciID, config.DriverVfioPci, true)
+		err := config.SwapDriver(d.log, d.intf.State.PciID, config.DriverVfioPci, true)
 		if err != nil {
-			return errors.Wrapf(err, "Couldnt swap %s to vfio_pci", d.conf.PciID)
+			return errors.Wrapf(err, "Couldnt swap %s to vfio_pci", d.intf.State.PciID)
 		}
 	}
 	return nil
 }
 
-func (d *Vmxnet3Driver) RestoreLinux(allInterfacesPhysical bool) {
-	if d.conf.PciID != "" && d.conf.Driver != "" {
-		err := utils.SwapDriver(d.conf.PciID, d.conf.Driver, true)
+func (d *Vmxnet3Driver) RestoreLinux() {
+	if d.intf.State.PciID != "" && d.intf.State.Driver != "" {
+		err := config.SwapDriver(d.log, d.intf.State.PciID, d.intf.State.Driver, true)
 		if err != nil {
-			log.Warnf("Error swapping back driver to %s for %s: %v", d.conf.Driver, d.conf.PciID, err)
+			d.log.Warnf("Error swapping back driver to %s for %s: %v", d.intf.State.Driver, d.intf.State.PciID, err)
 		}
 	}
 
-	if !d.conf.IsUp {
+	if !d.intf.State.IsUp {
 		return
 	}
 	// This assumes the link has kept the same name after the rebind.
 	// It should be always true on systemd based distros
-	link, err := utils.SafeSetInterfaceUpByName(d.spec.InterfaceName)
+	link, err := config.SafeSetInterfaceUpByName(d.intf.Spec.InterfaceName)
 	if err != nil {
-		log.Warnf("Error setting %s up: %v", d.spec.InterfaceName, err)
+		d.log.Warnf("Error setting %s up: %v", d.intf.Spec.InterfaceName, err)
 		return
 	}
 
@@ -86,9 +86,9 @@ func (d *Vmxnet3Driver) RestoreLinux(allInterfacesPhysical bool) {
 	d.restoreLinuxIfConf(link)
 
 	if d.params.InitialVfioEnableUnsafeNoIommuMode == config.VfioUnsafeNoIommuModeNO {
-		err = utils.SetVfioEnableUnsafeNoIommuMode(config.VfioUnsafeNoIommuModeNO)
+		err = config.SetVfioEnableUnsafeNoIommuMode(config.VfioUnsafeNoIommuModeNO)
 		if err != nil {
-			log.Errorf("Vmxnet3 restore error %s", err)
+			d.log.Errorf("Vmxnet3 restore error %s", err)
 		}
 	}
 }
@@ -97,28 +97,32 @@ func (d *Vmxnet3Driver) CreateMainVppInterface(vpp *vpplink.VppLink, vppPid int,
 	intf := types.Vmxnet3Interface{
 		GenericVppInterface: d.getGenericVppInterface(),
 		EnableGso:           *config.GetCalicoVppDebug().GSOEnabled,
-		PciID:               d.conf.PciID,
+		PciID:               d.intf.State.PciID,
 	}
 	swIfIndex, err := vpp.CreateVmxnet3(&intf)
 	if err != nil {
 		return errors.Wrapf(err, "Error creating Vmxnet3 interface")
 	}
 
-	log.Infof("Created Vmxnet3 interface %d", swIfIndex)
+	d.log.Infof("Created Vmxnet3 interface %d", swIfIndex)
 
-	d.spec.SwIfIndex = swIfIndex
-	err = d.TagMainInterface(vpp, swIfIndex, d.spec.InterfaceName)
+	d.intf.Spec.SwIfIndex = swIfIndex
+	err = d.TagMainInterface(vpp, swIfIndex, d.intf.Spec.InterfaceName)
 	if err != nil {
 		return err
 	}
 	return nil
 }
 
-func NewVmxnet3Driver(params *config.VppManagerParams, conf *config.LinuxInterfaceState, spec *config.UplinkInterfaceSpec) *Vmxnet3Driver {
-	d := &Vmxnet3Driver{}
-	d.name = NativeDriverVmxnet3
-	d.conf = conf
-	d.params = params
-	d.spec = spec
-	return d
+func NewVmxnet3Driver(params *vppmanagerparams.VppManagerParams,
+	intf *vppmanagerparams.VppManagerInterface,
+	log *logrus.Entry) *Vmxnet3Driver {
+	return &Vmxnet3Driver{
+		UplinkDriverData: UplinkDriverData{
+			name:   NativeDriverVmxnet3,
+			intf:   intf,
+			params: params,
+			log:    log,
+		},
+	}
 }
